Brought up to date.

+++ class-app/scripts/setup.js
@@ -5,12 +5,6 @@ import process from 'node:process';
 import path from 'node:path';
 import fs from 'fs-extra';
 
-/*
- * TODOs
- *
- * What if someone passes "lab-01" instead of "1" or "01"
- */
-
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
@@ -20,14 +14,19 @@ let labsParentPrefix = '../../labs';
 let labsLocalPrefix = '../labs';
 let backupPrefix = `../tmp/backup-${Date.now()}`;
 
-async function clean() {
+/**
+ * @param {string} directory
+ * @param {boolean} [relative=false]
+ */
+async function clean(directory, relative = false) {
 	try {
-		let target = path.resolve(__dirname, labsLocalPrefix);
+		let target = directory;
+		if (relative) target = path.resolve(__dirname, directory);
 		let result = await fs.emptyDir(target);
 		console.log(`Successfully cleaned ${target}`);
 		return result;
 	} catch (err) {
-		console.error(`clean(): ${err.message}`);
+		console.error(`clean(${directory}): ${err.message}`);
 		throw err;
 	}
 }
@@ -36,10 +35,15 @@ async function backup() {
 	let src = path.resolve(__dirname, labsLocalPrefix);
 	let dest = path.resolve(__dirname, backupPrefix);
 	try {
+		let srcExists = await fs.exists(src);
+		if (!srcExists) {
+			console.warn('Nothing to back up!');
+			return;
+		}
 		await fs.copy(src, dest, {
 			filter: (src) => !src.includes(finishedFolderName),
 		});
-		await clean();
+		await clean(src);
 		console.log(`Successfully backed up the labs folder to ${dest}`);
 	} catch (err) {
 		console.error(`backup(): ${err.message}`);
@@ -76,10 +80,31 @@ async function finish(labNumber) {
 		let labsFromExists = await fs.exists(labsFrom);
 		if (!labsFromExists) throw new Error(`Could not find lab ${labNumber}`);
 		let labsTo = path.resolve(__dirname, labsLocalPrefix, finishedFolderName, labDir);
+		await clean(labsTo);
 		await fs.copy(labsFrom, labsTo);
 		console.log(`Successfully copied the solution for ${labDir}`);
 	} catch (err) {
 		console.error(`finish(): ${err.message}`);
+	}
+}
+
+async function save(labNumber, beginFinish) {
+	let labDir = `lab-${labNumber}`;
+	let labsTo = path.resolve(
+		__dirname,
+		labsParentPrefix,
+		beginFinish === 'begin' ? beginFolderName : finishedFolderName,
+		labDir,
+	);
+	let labsFrom = path.resolve(__dirname, labsLocalPrefix);
+	try {
+		await clean(labsTo);
+		await fs.copy(labsFrom, labsTo, {
+			filter: (src) => !(src.includes('.gitkeep') || src.includes('finished')),
+		});
+		console.log(`Successfully saved files to ${labsTo}`);
+	} catch (err) {
+		console.error(`save(): ${err.message}`);
 	}
 }
 
@@ -100,6 +125,7 @@ const operations = {
 	begin,
 	clean,
 	finish,
+	save,
 	default: defaultOperation,
 };
 
@@ -107,8 +133,7 @@ let [operation, labNumber] = process.argv.slice(2);
 if (labNumber?.length === 1) labNumber = '0' + labNumber;
 
 // If the first argument is a number, do the default action
-if (Object.keys(operations)
-	.includes(operation)) {
+if (Object.keys(operations).includes(operation)) {
 	operations[operation](labNumber);
 } else if (!isNaN(Number(operation))) {
 	// Operation is actually a lab number
